+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Go to IAPT result page when GP has been selected from typeahead
 - Remove `no suggestions` display for zero typeahead results
 - Show 5 results in typeahead
+- Do not search postcode in typeahead
 
 0.2.0 / 2018-07-23
 ==================