+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Show 5 results in typeahead
 - Do not search postcode in typeahead
 - Scroll input to top for smaller screen sizes
+- Add Meta tags for WebTrends
 
 0.2.0 / 2018-07-23
 ==================